SMP/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
multidict==6.0.5
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==3.12.3
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
sqlparse==0.5.1
twilio==9.2.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.7
